requirements.txt
python-dotenv==1.0.0
pika==1.3.2
redis==5.0.1
requests==2.31.0
pandas==2.1.4 